Require PHP >5.4.15
The version of PHP listed in the composer.json was not the same
as the version of PHP installed on the Vagrant box. This is open
for discussion, obviously, but I would like this project to push
anyone wanting to use it to upgrade from 5.2 or 5.3 to 5.4 (or
possibly 5.5 depending on timeline).