Adding throwaway tests so the build will pass
Setting up Travis to run before having any actual code was not a
brilliant idea. There is now a simple test that verifies true is
true that will be removed once we have a real test added.

I also modified how to run the tests so that tests will be run on
the vagrant box rather than on the host machine.