Tweaks for Travis build
The Travis build was having issues because it was a) using an
outdated global install of composer and b) I had a requirement of
PHP v5.4.15 which is a problem on the Travis system or within
composer itself. I bumped the required version of PHP down to 5.4.0
to hopefully fix the build.