Integrating assetic and Twitter Bootstrap
Decided to use the Assetic project to handle asset management.
Integrated that using the Silex-Assetic project to concat all CSS
and JS resources into the public docroot. I also want to use LESS
for this project so I included LESS processing in the Assetic
setup.

Additionally I moved the docroot for the webserver from a synced
directory to the /vagrant directory which is synced by default.